requirements.txt
# LLM and Orchestration
langchain>=0.1.0,<0.2.0
langgraph>=0.0.20,<0.1.0
anthropic>=0.8.0,<1.0.0
openai>=1.0.0,<2.0.0

# Graph Database
neo4j>=5.14.0,<6.0.0
networkx>=3.2,<4.0

# Vector Store
chromadb>=0.4.18,<0.5.0

# Caching
redis>=5.0.0,<6.0.0

# NLP and Entity Extraction
spacy>=3.7.0,<4.0.0
transformers>=4.35.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0

# API Framework
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
websockets>=12.0,<13.0

# Data Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0

# Document Processing
pypdf>=3.17.0,<4.0.0
python-docx>=1.1.0,<2.0.0
markdown>=3.5.0,<4.0.0

# Utilities
tenacity>=8.2.0,<9.0.0
tiktoken>=0.5.0,<1.0.0
